requirements.txt
anthropic
httpx
supabase
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
openai
